+++ custom_scaler.py
@@ -1,0 +1,11 @@
+class CustomScaler:
+        def fit(self, X):
+            self.mean_ = np.mean(X, axis=0)
+            self.std_ = np.std(X, axis=0)
+            return self
+
+        def transform(self, X):
+            return (X - self.mean_) / self.std_
+
+        def fit_transform(self, X):
+            return self.fit(X).transform(X)
