+++ main.py
@@ -1,151 +1,9 @@
 import numpy as np
 from sklearn.datasets import load_iris
 from sklearn.model_selection import train_test_split
-from sklearn.preprocessing import StandardScaler
-class CustomScaler:
-        def fit(self, X):
-            self.mean_ = np.mean(X, axis=0)
-            self.std_ = np.std(X, axis=0)
-            return self
+from custom_scaler import CustomScaler 
+from multilayer_perceptron import MultilayerPerceptron  
 
-        def transform(self, X):
-            return (X - self.mean_) / self.std_
-
-        def fit_transform(self, X):
-            return self.fit(X).transform(X)
-        
-class MultilayerPerceptron:
-    def __init__(self, layer_sizes, learning_rate=0.01, n_iterations=1000):
-        """
-        Initialize the Multilayer Perceptron (MLP) with specified layer sizes, learning rate, and iterations.
-        
-        Parameters:
-        layer_sizes (list of int): Sizes of each layer, including input, hidden, and output layers.
-        learning_rate (float): The step size for weight updates.
-        n_iterations (int): The number of times to iterate over the training data.
-        
-        Initialize weights and biases for each layer.
-        """
-        # Initialize learning rate, iterations, and layer sizes.
-        # Initialize weights and biases for each layer randomly (use small values).
-        pass
-
-    def _initialize_weights(self, layer_sizes):
-        """
-        Initialize the weights and biases for the MLP layers.
-        
-        Parameters:
-        layer_sizes (list of int): Number of neurons in each layer (input, hidden, output).
-        
-        Initialize weights and biases as lists of numpy arrays where:
-        - Weights between layer `i` and `i+1` are of shape (layer_sizes[i], layer_sizes[i+1])
-        - Biases for layer `i+1` are of shape (layer_sizes[i+1],)
-        """
-        # Randomly initialize weights and biases for each layer.
-        pass
-
-    def _activation_function(self, z, activation="sigmoid"):
-        """
-        Apply activation function (e.g., sigmoid or ReLU) to the input z.
-        
-        Parameters:
-        z (numpy.ndarray): Linear combination of inputs and weights (pre-activation).
-        activation (str): The type of activation function to apply ("sigmoid", "relu").
-        
-        Returns:
-        numpy.ndarray: Activated output.
-        """
-        # Implement the sigmoid or ReLU activation function based on the parameter `activation`.
-        pass
-
-    def _activation_derivative(self, z, activation="sigmoid"):
-        """
-        Compute the derivative of the activation function for backpropagation.
-        
-        Parameters:
-        z (numpy.ndarray): Linear combination of inputs and weights (pre-activation).
-        activation (str): The type of activation function to compute the derivative for ("sigmoid", "relu").
-        
-        Returns:
-        numpy.ndarray: The derivative of the activation function with respect to z.
-        """
-        # Implement the derivative of the activation function for backpropagation.
-        pass
-
-    def _forward_propagation(self, X):
-        """
-        Perform forward propagation through the network.
-        
-        Parameters:
-        X (numpy.ndarray): Input data, shape (n_samples, n_features).
-        
-        Returns:
-        list: A list of activations for each layer, including input and output.
-        """
-        # Perform the forward pass through each layer, calculating the linear combination
-        # and applying the activation function. Store activations for backpropagation.
-        pass
-
-    def _backward_propagation(self, X, y, activations):
-        """
-        Perform backward propagation to calculate gradients and update weights.
-        
-        Parameters:
-        X (numpy.ndarray): Input data, shape (n_samples, n_features).
-        y (numpy.ndarray): True labels.
-        activations (list): List of activations from each layer during forward propagation.
-        
-        Update the weights and biases based on the gradients computed from backpropagation.
-        """
-        # Compute the error at the output layer and propagate it backward
-        # through the network to compute gradients for each weight and bias.
-        # Update the weights and biases using the learning rate.
-        pass
-
-    def fit(self, X, y):
-        """
-        Train the Multilayer Perceptron using forward and backward propagation.
-        
-        Parameters:
-        X (numpy.ndarray): Training data, shape (n_samples, n_features).
-        y (numpy.ndarray): True labels, shape (n_samples,).
-        
-        Perform multiple iterations of forward and backward propagation to adjust the weights.
-        """
-        # Loop through the training process for `n_iterations`.
-        # For each iteration, perform forward propagation and then backpropagation to adjust weights.
-        pass
-
-    def predict(self, X):
-        """
-        Make predictions using the trained MLP.
-        
-        Parameters:
-        X (numpy.ndarray): Data to predict, shape (n_samples, n_features).
-        
-        Returns:
-        numpy.ndarray: Predicted class labels (or probabilities, depending on output layer).
-        
-        Perform forward propagation and return the output of the final layer.
-        """
-        # Perform a forward pass and return the predicted class labels.
-        pass
-
-    def _loss(self, y_true, y_pred):
-        """
-        Calculate the loss function (e.g., cross-entropy loss or MSE).
-        
-        Parameters:
-        y_true (numpy.ndarray): True labels.
-        y_pred (numpy.ndarray): Predicted labels or probabilities.
-        
-        Returns:
-        float: Loss value.
-        """
-        # Implement the loss function (cross-entropy for classification, MSE for regression).
-        pass
-
-# Example usage (add real dataset in place of dummy data):
 if __name__ == "__main__":
     # Load the iris dataset
     iris = load_iris()
@@ -170,8 +28,3 @@ if __name__ == "__main__":
 
     # Print the predictions
     print(predictions)
-    # Define the dataset (X: features, y: labels)
-    # Initialize the MLP with a list of layer sizes (e.g., input layer, hidden layers, output layer)
-    # Train the model using the `fit` function
-    # Use the `predict` function to classify new data
-    pass
